\name{categoryUtilityFunction}
\alias{categoryUtilityFunction}
\title{Calculates the category utility function
Sums the differences between conditional and marginal probabilities
for each attribute}
\usage{
  categoryUtilityFunction(x)
}
\arguments{
  \item{x}{partition object}
}
\description{
  Calculates the category utility function Sums the
  differences between conditional and marginal
  probabilities for each attribute
}

