\docType{data}
\name{entropy}
\alias{entropy}
\title{Entropy for a vector.
Omits NA. If you consider NA as a value itself,
use is.na to deactivate it.}
\format{List of 3
 $ name   : chr "entropy"
 $ address:Class 'NativeSymbol' <externalptr> 
 $ package: NULL
 - attr(*, "class")= chr "NativeSymbolInfo"}
\usage{
  entropy
}
\arguments{
  \item{vector}{}
}
\value{
  total entropy for vector
}
\description{
  Entropy for a vector. Omits NA. If you consider NA as a
  value itself, use is.na to deactivate it.
}
\keyword{datasets}

