\name{expectedEntropy.pairwise}
\alias{expectedEntropy.pairwise}
\title{Expected entropy of Y given X, where Y is factor
Assumes independence of variables on Y}
\usage{
  expectedEntropy.pairwise(y, x)
}
\arguments{
  \item{y}{factor}

  \item{x}{factor to clustering}
}
\description{
  Expected entropy of Y given X, where Y is factor Assumes
  independence of variables on Y
}

