\name{expectedEntropy}
\alias{expectedEntropy}
\title{Expected entropy of Y given X, where Y is multivariate
Assumes independence of variables on Y}
\usage{
  expectedEntropy(x, cluster)
}
\arguments{
  \item{x}{data.frame}

  \item{cluster}{factor to clustering}
}
\description{
  Expected entropy of Y given X, where Y is multivariate
  Assumes independence of variables on Y
}

