\name{fit.point}
\alias{fit.point}
\title{Returns the cluster for which the point is better suited
Using expectedEntropy as criteria}
\usage{
  fit.point(xx, i)
}
\arguments{
  \item{xx}{partition object}

  \item{i}{number index}
}
\value{
  number of conglomerate
}
\description{
  Returns the cluster for which the point is better suited
  Using expectedEntropy as criteria
}

