\name{fitPoint}
\alias{fitPoint}
\title{Returns the cluster for which the point is better suited.
Using expectedEntropy as criteria}
\usage{
  fitPoint(xx, i)
}
\arguments{
  \item{xx}{partition object}

  \item{i}{number index}
}
\value{
  number of conglomerate
}
\description{
  Returns the cluster for which the point is better suited.
  Using expectedEntropy as criteria
}

