\name{hummingDistance}
\alias{hummingDistance}
\title{Distancia de humming entre vectores
Es m치s c칩modo utilizar daisy del paquete cluster, que aparte
entrega el valor estandarizado}
\usage{
  hummingDistance(x, y = NULL)
}
\description{
  Distancia de humming entre vectores Es m치s c칩modo
  utilizar daisy del paquete cluster, que aparte entrega el
  valor estandarizado
}

