\name{pointFittingProbability}
\alias{pointFittingProbability}
\title{Point fitting probability
Sum all the p for each attribute on a case}
\usage{
  pointFittingProbability(xx, first = 1,
    last = nrow(xx$data))
}
\arguments{
  \item{xx}{coolcat object}

  \item{first}{index of case to start}

  \item{last}{index of case to finnish}
}
\description{
  Point fitting probability Sum all the p for each
  attribute on a case
}

