\name{symmetricUncertaintyCoefficient}
\alias{symmetricUncertaintyCoefficient}
\title{Symmetrical measure of uncertainty
between two categorical variables}
\usage{
  symmetricUncertaintyCoefficient(x, y)
}
\arguments{
  \item{x}{factor}

  \item{y}{factor}
}
\value{
  real value
}
\description{
  Symmetrical measure of uncertainty between two
  categorical variables
}

