\name{symUncertaintyCoefficient}
\alias{symUncertaintyCoefficient}
\title{Symmetrical measure of uncertainty
between variables on a data.frame or
two categorical variables (factor).
Range: 0 (no dependence)- 1(complete dependence)
Removes NA from both factors}
\usage{
  symUncertaintyCoefficient(x, y = NULL)
}
\arguments{
  \item{x}{data.frame or first factor}

  \item{y}{second factor}
}
\value{
  real value
}
\description{
  Symmetrical measure of uncertainty between variables on a
  data.frame or two categorical variables (factor). Range:
  0 (no dependence)- 1(complete dependence) Removes NA from
  both factors
}

