\name{tableByFactor}
\alias{tableByFactor}
\title{Percentages or frequency for attribute on a data.frame, ordered by factor}
\usage{
  tableByFactor(x, f, values = "percentage")
}
\arguments{
  \item{x}{dataframe}

  \item{f}{factor which divide the data.frame}

  \item{values}{percentage or frequency}
}
\value{
  a list. Every dataframe with percentages has the name of
  the attribute
}
\description{
  Percentages or frequency for attribute on a data.frame,
  ordered by factor
}

