\name{categoryUtilityFunction}
\alias{categoryUtilityFunction}
\title{Calculates the category utility function
Sums the differences between conditional and marginal probabilities
for each attribute}
\usage{
  categoryUtilityFunction(d, g)
}
\arguments{
  \item{d}{data}

  \item{c}{clustering vector}
}
\description{
  Calculates the category utility function Sums the
  differences between conditional and marginal
  probabilities for each attribute
}

