\name{entropy}
\alias{entropy}
\title{Entropy for a vector.
Omits NA. If you consider NA as a value itself,
use is.na to deactivate it.}
\usage{
  entropy(x)
}
\arguments{
  \item{x}{factor}
}
\value{
  total entropy for vector
}
\description{
  Entropy for a vector. Omits NA. If you consider NA as a
  value itself, use is.na to deactivate it.
}

