\name{pointFittingProbability}
\alias{pointFittingProbability}
\title{Point fitting probability
Calculates the log of conditional probability for each case
given the clustering factor.}
\usage{
  pointFittingProbability(xx, indexes = 1:nrow(xx$data))
}
\arguments{
  \item{xx}{coolcat object}

  \item{first}{index of case to start}

  \item{last}{index of case to finnish}
}
\description{
  Point fitting probability Calculates the log of
  conditional probability for each case given the
  clustering factor.
}

