\name{pointFittingProbability}
\alias{pointFittingProbability}
\title{Point fitting probability
Sum all the p for each attribute on a case}
\usage{
  pointFittingProbability(xx, m, first = 1,
    last = nrow(xx$data))
}
\description{
  Point fitting probability Sum all the p for each
  attribute on a case
}

